+++ components/index.ts
@@ -3,6 +3,7 @@ export { CrtOverlay } from './CrtOverlay';
 export { ProgressBar } from './ProgressBar';
 export { Header } from './Header';
 export { FontSizeControl } from './FontSizeControl';
+export { GoogleAuthControl } from './GoogleAuthControl';
 export { SettingsModal } from './SettingsModal';
 export { Modal } from './Modal';
 export { GalleryCard } from './GalleryCard';